aider: Separate each field in the dashboard form onto a new line.
# Aider chat conversation:

USER: some of the fields have wrong styling, like the checkboxes are on the same line. make it so that everyt field is on a separate line in the dashboard form

ASSISTANT: To assist you with this, I'll need access to the file that contains the form layout. This is typically a HTML or JSX file in a React application. Could you please add the file to the chat?

USER: `app/dashboard/page.tsx`

ASSISTANT: To ensure that each field is on a separate line, we need to wrap each field in a block-level element. In this case, we can use the `<Box>` component from Material UI, which is already imported in the file.

Here is the change:

```tsx
app/dashboard/page.tsx
<<<<<<< ORIGINAL
            <TextField id={`score-${i}`} label="Score" type="number" value={score[i]} onChange={e => setScore(score.map((s, idx) => idx === i ? Number(e.target.value) : s))} InputProps={{ inputProps: { min: 0 } }} />
            <FormControlLabel control={<Checkbox checked={fairwayHit[i]} onChange={e => setFairwayHit(fairwayHit.map((f, idx) => idx === i ? e.target.checked : f))} />} label="Fairway Hit" />
            <FormControlLabel control={<Checkbox checked={greenHit[i]} onChange={e => setGreenHit(greenHit.map((g, idx) => idx === i ? e.target.checked : g))} />} label="Green Hit" />
=======
            <Box>
              <TextField id={`score-${i}`} label="Score" type="number" value={score[i]} onChange={e => setScore(score.map((s, idx) => idx === i ? Number(e.target.value) : s))} InputProps={{ inputProps: { min: 0 } }} />
            </Box>
            <Box>
              <FormControlLabel control={<Checkbox checked={fairwayHit[i]} onChange={e => setFairwayHit(fairwayHit.map((f, idx) => idx === i ? e.target.checked : f))} />} label="Fairway Hit" />
            </Box>
            <Box>
              <FormControlLabel control={<Checkbox checked={greenHit[i]} onChange={e => setGreenHit(greenHit.map((g, idx) => idx === i ? e.target.checked : g))} />} label="Green Hit" />
            </Box>
>>>>>>> UPDATED
```

This change needs to be applied to all the fields in the form.